pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cmdbsyncer-inventory"
version = "0.1.3"
authors = [
    {name = "Bastian Kuhn", email = "bastian.kuhn@kuhn-ruess.de"},
]
description = "Ansible Inventory Plugin for CMDBSyncer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8", 
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Networking",
]
keywords = ["ansible", "inventory", "cmdbsyncer", "automation", "infrastructure"]
dependencies = [
    "ansible-core>=2.12",
    "requests>=2.25.0",
]

[project.urls]
Homepage = "https://github.com/kuhn-ruess/cmdbsyncer-inventory"
Documentation = "https://github.com/kuhn-ruess/cmdbsyncer-inventory/wiki"
Repository = "https://github.com/kuhn-ruess/cmdbsyncer-inventory.git"
Issues = "https://github.com/kuhn-ruess/cmdbsyncer-inventory/issues"

[project.entry-points."ansible.inventory"]
cmdbsyncer_inventory = "cmdbsyncer_inventory.inventory_plugins.cmdbsyncer_inventory:InventoryModule"
[project.entry-points."console_scripts"]
cmdbsyncer-install-plugin = "cmdbsyncer_inventory:install_plugin"

[tool.setuptools.packages.find]
where = ["."]
include = ["cmdbsyncer_inventory*"]

[tool.setuptools.package-data]
cmdbsyncer_inventory = ["*.py"]